requirements.txt
streamlit==1.38.0
fastapi==0.114.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
openpyxl==3.1.5
requests==2.32.3
pydub==0.25.1
# torch==2.3.1
# torchaudio==2.3.1
openai-whisper==20231117
silero-vad==5.1.2
librosa==0.10.2
soundfile==0.12.1
noisereduce==3.0.2
opencc-python-reimplemented==0.1.7
